requirements.txt
numpy==2.2.1
Pillow==11.1.0
streamlit==1.41.1
streamlit==1.30.0
tensorflow==2.10.0
tensorflow==2.17.0
tensorflow_intel==2.17.0
